refactor: move query iterator into module
